+++ paymentsheet-example/src/androidTest/java/com/stripe/android/lpm/TestBancontact.kt
@@ -2,6 +2,10 @@ package com.stripe.android.lpm
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.stripe.android.BaseLpmTest
+import com.stripe.android.test.core.AuthorizeAction
+import com.stripe.android.test.core.Automatic
+import com.stripe.android.test.core.DelayedPMs
+import com.stripe.android.test.core.IntentType
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -15,6 +19,29 @@ internal class TestBancontact : BaseLpmTest() {
     fun testBancontact() {
         testDriver.confirmNewOrGuestComplete(
             testParameters = bancontact,
+        )
+    }
+
+    @Test
+    fun testBancontactSfu() {
+        testDriver.confirmNewOrGuestComplete(
+            testParameters = bancontact.copy(
+                delayed = DelayedPMs.On,
+                automatic = Automatic.On,
+                intentType = IntentType.PayWithSetup,
+            ),
+        )
+    }
+
+    @Test
+    fun testBancontactSetup() {
+        testDriver.confirmNewOrGuestComplete(
+            testParameters = bancontact.copy(
+                delayed = DelayedPMs.On,
+                automatic = Automatic.On,
+                intentType = IntentType.Setup,
+                authorizationAction = AuthorizeAction.AuthorizeSetup,
+            ),
         )
     }